+++ meta/CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to the Orbital Trading project will be documented in this file by the user, only. Gemini is never to modify or provide updates to this file, and must always treat this file as read-only. The user will always manually update this file.
 
 
+# [37.17] - 2026-3-24
+-  added number resizing for qty input on market screen
+-  increased repair costs by 31.8%
+-  increast hull deterioration rate by 5%
+
 # [37.14] - 2026-3-24
 -  market behavior bug fixes, balance, and polish
 -  fixed price graph regressions